Simplified the implementation of Working_With_Lists/main_list_subset_generator.py
This Python script prompts the user to input a space delimited set of items,
it parses them simply and then instantiates an instance of ListSubsetGenerator
to generate all of the subsets of that list. It generates each subset one at
a time and prints it out, it prints f"{i}: {subset}" where i is an index,
starting with 1 and going to 2**n - 1 until all of the subsets are generated
and enumerated.

In my previous implementation I had the print statement on two lines,
first:

        print(f"{i}: ", end='')

Next:

        print(subset)

This works. In this revision I simplified it into one line:

        print(f"{i}: {subset}")